move components to centralize location
