+++ app/(navWithSide)/projects/[projectId]/board/page.tsx
@@ -1,10 +1,10 @@
 "use client";
 import Header from "@/components/ui/Header";
-import TaskCard from "./_components/TaskCard";
+import TaskCard from "@/components/task/TaskCard";
 import { Select, SelectItem } from "@heroui/select";
 import { Button } from "@heroui/button";
 import { useDisclosure } from "@heroui/modal";
-import SprintGoalModal from "./_components/SprintGoalModal";
+import SprintGoalModal from "@/components/task/SprintGoalModal";
 
 export default function BoardPage() {
   const { isOpen, onOpen, onOpenChange } = useDisclosure();
